fix: address all reviewer comments on PR #1876
- blockchain: rename BlockchainStorageV4→V5, add legacy V4 for backward compat, bump FILE_VERSION to 5
- blockchain: extract cbe_token_id() helper to eliminate repeated CBE id derivation
- exchange: clarify SOV/USDC helpers are test-only, not a live pair
- onramp: derive VWAP_WINDOW_BLOCKS from TARGET_BLOCK_TIME_SECONDS=10 (was hardcoded for 6s)
- onramp: fix record_trade doc to accurately describe MAX_TRADE_HISTORY cap
- consensus: use try_write() for ZK proof task to never block consensus reads
- runtime: return error (not warning) when wire_blockchain_to_consensus fails
- bonding_curve: fix u64 overflow in reserve split calc; fix comment 20/80→40/60

+++ zhtp/src/api/handlers/bonding_curve/api_v1.rs
@@ -296,8 +296,8 @@ impl BondingCurveApiHandler {
             ));
         }
 
-        // Calculate 20/80 split using the canonical constant
-        let to_reserve = req.sov_amount * RESERVE_SPLIT_NUMERATOR / RESERVE_SPLIT_DENOMINATOR;
+        // Calculate 40/60 split (RESERVE_SPLIT_NUMERATOR/DENOMINATOR = 2/5) using u128 to prevent overflow
+        let to_reserve = (req.sov_amount as u128 * RESERVE_SPLIT_NUMERATOR as u128 / RESERVE_SPLIT_DENOMINATOR as u128) as u64;
         let to_treasury = req.sov_amount - to_reserve;
 
         // Calculate CBE output using the contract's integer math
